Crate dual_quaternion

A simple and type agnostic dual-quaternion math library designed for reexporting

Functions

add
Adds two dual-quaternions

conj
Returns the dual-quaternion conjugate

dot
Dot product of two dual-quaternions

from_rotation_and_translation
Construct a dual-quaternion from separate rotation and translation components

get_rotation
Extracts rotation component from a dual-quaternion

get_translation
Extracts translation component from a dual-quaternion

id
Constructs identity dual-quaternion, representing no rotation or translation.

mul
Multiplies two dual-quaternions

normalize
Normalizes a dual-quaternion

scale
Scales a dual-quaternion (element-wise) by a scalar

Type Definitions

DualQuaternion
A dual-quaternion consists of a real component and a dual component, and can be used to represent both rotation and translation